actualizar crud de usuarios y corrección de controllers

+++ src/main/java/com/levelup/service/RoleService.java
@@ -38,6 +38,9 @@ public class RoleService {
     }
 
     public void deleteRole(Long id){
+        if (!roleRepo.existsById(id)) {
+            throw new IllegalArgumentException("Role not found");
+        }
         roleRepo.deleteById(id);
     }
 }
